multiple trackers working properly
